ci: add test and GitHub Actions workflow for CI
Added a GitHub Actions workflow to run tests on multiple Emacs versions.
Configured the workflow to trigger on push and pull request events. The
Makefile was also added to define test commands and dependencies. Updated
test files to ensure proper setup and teardown of the git repository for
consistent test results.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>